pyproject.toml
[project]
name = "yourproj"
version = "0.3.0"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.26",
    "scikit-learn>=1.3",
    "pyyaml>=6.0.2",
    "matplotlib>=3.7",
    "ipython>=8.0",
    "ipykernel",
    "nbformat",
    "nbconvert",
    "pre-commit",
    "ruff",
    "black",
    "pytest",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"


[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
